requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
fonttools==4.58.0
h11==0.16.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-louvain==0.16
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
